requirements.txt
fastapi
pydantic
httpx
python-dotenv
google-generativeaid
gradio
uvicorn
dash
